requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
ollama==0.1.9
coqui-tts==0.27.5
pydantic==2.5.0
python-multipart==0.0.6
streamlit==1.28.1
requests==2.31.0
numpy==1.26.4
scipy==1.11.4
whisper==1.1.10
torch>=2.0.0
torchaudio
deep-translator==1.11.4
f5-tts
transformers>=4.30.0
gTTS>=2.3.2
python-telegram-bot==20.0
pydub==0.25.1
